+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features
 
 - Customizable fragment lifecycle breadcrumbs ([#2299](https://github.com/getsentry/sentry-java/pull/2299))
+- Provide hook for Jetpack Compose navigation instrumentation ([#2320](https://github.com/getsentry/sentry-java/pull/2320))
 
 ## 6.6.0